+++ mock/cpp_mock_scenarios/src/metrics/traveled_distance.cpp
@@ -54,9 +54,9 @@ private:
 
   void onInitialize() override
   {
-    api_.spawn(false, "ego", getVehicleParameters());
-    api_.setEntityStatus(
-      "ego", traffic_simulator::helper::constructLaneletPose(34741, 0, 0),
+    api_.spawn(
+      false, "ego", getVehicleParameters(),
+      traffic_simulator::helper::constructLaneletPose(34741, 0, 0),
       traffic_simulator::helper::constructActionStatus(3));
     api_.setTargetSpeed("ego", 3, true);
     api_.addMetric<metrics::TraveledDistanceMetric>("ego_traveled_distance", "ego");