+++ simulation/traffic_simulator/include/traffic_simulator/entity/misc_object_entity.hpp
@@ -26,7 +26,8 @@ class MiscObjectEntity : public EntityBase
 {
 public:
   MiscObjectEntity(
-    const std::string & name, const openscenario_msgs::msg::MiscObjectParameters & params);
+    const std::string & name, const openscenario_msgs::msg::MiscObjectParameters & params,
+    const openscenario_msgs::msg::EntityStatus & status);
   void onUpdate(double, double) override;
   auto getBoundingBox() const -> const openscenario_msgs::msg::BoundingBox override;
   auto getCurrentAction() const -> const std::string override;