+++ simulation/traffic_simulator/include/traffic_simulator/entity/vehicle_entity.hpp
@@ -40,7 +40,8 @@ class VehicleEntity : public EntityBase
 {
 public:
   VehicleEntity(
-    const std::string & name, const openscenario_msgs::msg::VehicleParameters & parameters);
+    const std::string & name, const openscenario_msgs::msg::VehicleParameters & parameters,
+    const openscenario_msgs::msg::EntityStatus & status);
 
   ~VehicleEntity() override = default;
 
@@ -105,10 +106,7 @@ public:
     try {
       return tree_ptr_->getWaypoints();
     } catch (const std::runtime_error & e) {
-      if (!status_) {
-        THROW_SIMULATION_ERROR("Entity : ", name, " status is empty.");
-      }
-      if (status_ && status_->lanelet_pose_valid == false) {
+      if (status_.lanelet_pose_valid == false) {
         THROW_SIMULATION_ERROR(
           "Failed to calculate waypoints in NPC logics, please check Entity : ", name,
           " is in a lane coordinate.");
@@ -129,13 +127,10 @@ public:
 
   std::vector<std::int64_t> getRouteLanelets(double horizon = 100) override
   {
-    if (!status_) {
+    if (status_.lanelet_pose_valid) {
       return {};
     }
-    if (!status_->lanelet_pose_valid) {
-      return {};
-    }
-    return route_planner_ptr_->getRouteLanelets(status_->lanelet_pose, horizon);
+    return route_planner_ptr_->getRouteLanelets(status_.lanelet_pose, horizon);
   }
 
 private: