+++ simulation/traffic_simulator/src/entity/ego_entity.cpp
@@ -168,8 +168,9 @@ EgoEntity::EgoEntity(
   const std::string & name,             //
   const Configuration & configuration,  //
   const double step_time,               //
-  const openscenario_msgs::msg::VehicleParameters & parameters)
-: VehicleEntity(name, parameters),
+  const openscenario_msgs::msg::VehicleParameters & parameters,
+  const openscenario_msgs::msg::EntityStatus & status)
+: VehicleEntity(name, parameters, status),
   autoware(makeAutoware(configuration)),
   vehicle_model_type_(getVehicleModelType()),
   vehicle_model_ptr_(makeSimulationModel(vehicle_model_type_, step_time, parameters))
@@ -414,10 +415,8 @@ void EgoEntity::requestLaneChange(const std::int64_t)
     "everything but their destination");
 }
 
-bool EgoEntity::setStatus(const openscenario_msgs::msg::EntityStatus & status)
+void EgoEntity::setStatus(const openscenario_msgs::msg::EntityStatus &)
 {
-  const bool success = VehicleEntity::setStatus(status);  // NOTE: setStatus always succeeds.
-
   const auto current_pose = getStatus().pose;
 
   if (autoware->initialized()) {
@@ -434,8 +433,6 @@ bool EgoEntity::setStatus(const openscenario_msgs::msg::EntityStatus & status)
   if (!initial_pose_) {
     initial_pose_ = current_pose;
   }
-
-  return success;
 }
 
 void EgoEntity::setTargetSpeed(double value, bool)