+++ simulation/traffic_simulator/src/entity/pedestrian_entity.cpp
@@ -25,8 +25,9 @@ namespace traffic_simulator
 namespace entity
 {
 PedestrianEntity::PedestrianEntity(
-  const std::string & name, const openscenario_msgs::msg::PedestrianParameters & params)
-: EntityBase(params.pedestrian_category, name), parameters(params)
+  const std::string & name, const openscenario_msgs::msg::PedestrianParameters & params,
+  const openscenario_msgs::msg::EntityStatus & status)
+: EntityBase(params.pedestrian_category, name, status), parameters(params)
 {
   entity_type_.type = openscenario_msgs::msg::EntityType::PEDESTRIAN;
   tree_ptr_ = std::make_shared<entity_behavior::pedestrian::BehaviorTree>();
@@ -37,13 +38,10 @@ void PedestrianEntity::requestAssignRoute(
   const std::vector<openscenario_msgs::msg::LaneletPose> & waypoints)
 {
   tree_ptr_->setRequest("follow_lane");
-  if (!status_) {
+  if (status_.lanelet_pose_valid) {
     return;
   }
-  if (!status_->lanelet_pose_valid) {
-    return;
-  }
-  route_planner_ptr_->getRouteLanelets(status_->lanelet_pose, waypoints);
+  route_planner_ptr_->getRouteLanelets(status_.lanelet_pose, waypoints);
 }
 
 void PedestrianEntity::requestAssignRoute(const std::vector<geometry_msgs::msg::Pose> & waypoints)
@@ -66,13 +64,10 @@ void PedestrianEntity::requestAcquirePosition(
   const openscenario_msgs::msg::LaneletPose & lanelet_pose)
 {
   tree_ptr_->setRequest("follow_lane");
-  if (!status_) {
+  if (status_.lanelet_pose_valid) {
     return;
   }
-  if (!status_->lanelet_pose_valid) {
-    return;
-  }
-  route_planner_ptr_->getRouteLanelets(status_->lanelet_pose, lanelet_pose);
+  route_planner_ptr_->getRouteLanelets(status_.lanelet_pose, lanelet_pose);
 }
 
 void PedestrianEntity::requestAcquirePosition(const geometry_msgs::msg::Pose & map_pose)
@@ -99,19 +94,16 @@ void PedestrianEntity::setTargetSpeed(double target_speed, bool continuous)
 void PedestrianEntity::onUpdate(double current_time, double step_time)
 {
   EntityBase::onUpdate(current_time, step_time);
-  if (!status_) {
-    return;
-  }
   if (current_time < 0) {
     updateEntityStatusTimestamp(current_time);
   } else {
     tree_ptr_->setValueToBlackBoard("other_entity_status", other_status_);
     tree_ptr_->setValueToBlackBoard("entity_type_list", entity_type_list_);
-    tree_ptr_->setValueToBlackBoard("entity_status", status_.get());
-    target_speed_planner_.update(status_->action_status.twist.linear.x);
+    tree_ptr_->setValueToBlackBoard("entity_status", status_);
+    target_speed_planner_.update(status_.action_status.twist.linear.x);
     tree_ptr_->setValueToBlackBoard("target_speed", target_speed_planner_.getTargetSpeed());
-    if (status_->lanelet_pose_valid) {
-      auto route = route_planner_ptr_->getRouteLanelets(status_->lanelet_pose);
+    if (status_.lanelet_pose_valid) {
+      auto route = route_planner_ptr_->getRouteLanelets(status_.lanelet_pose);
       tree_ptr_->setValueToBlackBoard("route_lanelets", route);
     } else {
       std::vector<std::int64_t> empty = {};
@@ -128,13 +120,9 @@ void PedestrianEntity::onUpdate(double current_time, double step_time)
         return;
       }
     }
-    if (!status_) {
-      linear_jerk_ = 0;
-    } else {
-      linear_jerk_ =
-        (status_updated.action_status.accel.linear.x - status_->action_status.accel.linear.x) /
-        step_time;
-    }
+    linear_jerk_ =
+      (status_updated.action_status.accel.linear.x - status_.action_status.accel.linear.x) /
+      step_time;
     setStatus(status_updated);
     updateStandStillDuration(step_time);
   }