+++ test_runner/scenario_test_runner/launch/replay_bag.launch.py
@@ -31,7 +31,7 @@ from pathlib import Path
 
 def generate_launch_description():
     rviz_config_dir = os.path.join(
-        get_package_share_directory("scenario_test_runner"),
+        get_package_share_directory("traffic_simulator"),
         "config/scenario_simulator_v2.rviz",
     )
     print(rviz_config_dir)