+++ docs/ReleaseNotes.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## Difference between the latest release and master
+- Remove entity status which is empty. ([link](https://github.com/tier4/scenario_simulator_v2/pull/543)).
 - Update syntaxes of `openscenario_interpreter` to be compiled separately ([link](https://github.com/tier4/scenario_simulator_v2/pull/532)).
 
 ## Version 0.5.1